Add files via upload
This Repository houses a simple Hand written numeric digit recognition algorithm with an overall accuracy of 86% to classify/detect/predict any handwritten digit shown to it. This project is good to start with your neural network/deep learning journey, as it makes the use of simple linear algebra i.e matrix and some knowledge of basic 1D NumPy.